requirements.txt
flake8
python-keystoneclient
python-novaclient
python-neutronclient
python-cinderclient
google-api-python-client
python-dateutil
